change the page url mappings
