--- pages/collection/update/[urlPath].tsx
@@ -1,1 +1,0 @@
-// @TODO
